Separate table, log & registrations components Co-authored-by: Domenico Gemoli <domenicogemoli@gmail.com>
