+++ e2e/portalicious/tests/SendMessage/SendTemplatedMessage.spec.ts
@@ -8,6 +8,7 @@ import { registrationsPV } from '@121-service/test/registrations/pagination/pagi
 
 import BasePage from '@121-e2e/portalicious/pages/BasePage';
 import LoginPage from '@121-e2e/portalicious/pages/LoginPage';
+import RegistrationActivityLogPage from '@121-e2e/portalicious/pages/RegistrationActivityLogPage';
 import RegistrationsPage from '@121-e2e/portalicious/pages/RegistrationsPage';
 import TableComponent from '@121-e2e/portalicious/pages/TableComponent';
 
@@ -37,6 +38,7 @@ test('[31076] Send templated message', async ({ page }) => {
   const basePage = new BasePage(page);
   const registrations = new RegistrationsPage(page);
   const table = new TableComponent(page);
+  const activityLog = new RegistrationActivityLogPage(page);
 
   const projectTitle = 'NLRC Direct Digital Aid Program (PV)';
 
@@ -45,7 +47,7 @@ test('[31076] Send templated message', async ({ page }) => {
   });
 
   await test.step('Send templated message', async () => {
-    await registrations.selectAllRegistrations();
+    await table.selectAll();
     await registrations.selectBulkAction('Message');
     await registrations.selectTemplatedMessage('Include');
     await registrations.clickContinueToPreview();
@@ -55,7 +57,7 @@ test('[31076] Send templated message', async ({ page }) => {
     await registrations.sendMessage();
 
     await registrations.validateToastMessage(sendingMessageToast);
-    await table.selectRandomRegistration();
-    await registrations.validateLastMessageSent(includeMessageTemplate);
+    await registrations.selectRandomRegistration();
+    await activityLog.validateLastMessageSent(includeMessageTemplate);
   });
 });
